+++ Course 2/Algorithms_and_Data_Structures/Labs2-3/src/graph/Graph.h
@@ -57,7 +57,7 @@ public:
     bool haveCycle(Edge edge);
 
 private:
-    Graph(int verticesAmount);
+    Graph(size_t verticesAmount);
 
     void DepthFirstSearch(bool marked[], int currentIndex);
